requirements.txt
streamlit>=1.38.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
plotly>=5.24.0
requests>=2.32.0
groq>=0.11.0
python-dotenv>=1.0.1
openpyxl>=3.1.5
pytest>=8.0.0
